Add proxy differ
